Improve handling of unknown isolation groups (#7143)
Include an explicit metric for unknown isolation groups and eagerly abandon isolation. This metric is useful for distinguishing between isolation groups that only exist in another region and isolation groups that exist but don't have pollers. This is another place we'll need to change should we switch to fully dynamic isolation group support.

Additionally update task_reader to defensively check whether a buffer exists before trying to dispatch a task to it. Dispatching to a nil channel blocks forever.