Use a target-arch-compatible (and more up to date) version of dockerize (#7149)
Hopefully fixes #7093 in a simpler way.

Tested with `docker run ... dockerize --help` on two machines with arm64 and amd64 and it seems to work, though the `ARG TARGETARCH` appears to require `docker buildx` or manual entry.  That's a bit awkward for my test-machines (no buildx), but seems likely fine as github actions uses that for build-and-push.

And checked the final in-container docker.yml file that it generates, and it both ran fine and looks essentially identical.  
So this seems safe enough.
If we have issues, people can downgrade and we can revert (though revert will not restore a working arm image for new attempts.  but that's also broken now).