Added shard not found error (#7186)
What changed?
Added a shard not found error to the proto def, and mappers.
We now return this specific error when we get a lookup for a shard we do not know.

Why?
This error is an expected user error so we should treat it as such.

How did you test it?
Unit tests and local runs

Potential risks

Release notes

Documentation Changes