Don't attempt to dispatch expired activities (#7204)
We check whether an activity is expired before putting it into the buffer, but once it's there the only way we will find out that it has expired is from the Record(Decision/Activity)TaskStarted call returning an error. This is particularly problematic for TaskLists that have an Activity Rate Limit, since they'll only be able to remove tasks from the buffer according to that Rate Limit. They can enter a steady state where new non-expired tasks are added to the end of the buffer but they'll always expire by the time they actually are attempted.

Eagerly removing expired tasks ensures we can promptly dispatch relevant ones.