Support for creating ephemeral shards (#7209)
What changed?
We now support ephemeral shards

Shards are created if they do not already exist.
An executors shards are no longer deleted when the executor disappears
We now load the set of shards from the database for ephemeral shards
Future PRs will:

Support removing shards when they are no longer needed
Support the executor library to heartbeat immediately if an unknown shard is requested (so we do not have to wait for the next heartbeat to process shard requests)
Why?
We need to support ephemeral shards for use cases where the set of shard is changing (e.g. Cadence Matching)

How did you test it?

Potential risks

Release notes

Documentation Changes