Fixing flaky-tests in matching (#7213)
Multiple tests in matching were flaky because of race-condition:
 - we had to wait for engine.Stop() to finish since it calls finilizers
   we're checking with mocks
 - but we were calling it in goroutine w/o waiting

Instead, it is closer now to how code runs it -
runMembershipChangeLoop() is in goroutine, .Stop is called
synchronously.

Also renamed `subscribeToMembershipChanges` to `runMembershipChangeLoop`
to emphasize it is blocking a loop, not just "subscription".

In addition, removed time.Sleep - so now tests run for few ms, instead
of 3+ seconds.

Removed `TestMembershipSubscriptionShutdown` since it was 1:1 copying
`TestSubscriptionAndShutdown`.