Integration of removal of ephemeral shards in the executor library (#7216)
What changed?
The executor library now asks the processor for a struct of status information,
The new struct contains the status of the shard so the processor can signal the state of the shard

Why?
This enables the executors to tell if an ephemeral shard should be delete.

It also enables us to add more fields to the status in a backwards compatible way

How did you test it?
Unit tests

Potential risks
This is not backwards compatible, but we are still in the prototype phase so noone is implementing this interface, except our own test service.

Release notes

Documentation Changes