Make metric indexes unique (#7237)
Part cleanup due to discoveries in e.g. #7210, part prep-work for migrating metrics to almost literally any new setup (to make that easier and safer, by allowing int -> name lookups).

I *believe* this is a safe thing to do, as we don't seem to do anything fancy with the metric-def indexes.
But it's quite hard to verify because we don't have a unique type (fixed in #7238) and all the code here is rather value-oriented and fallback-y so it's hard to be totally confident that everything is handled.