[active-active] Fix failover version increment logic (#7246)
<!-- Describe what has changed in this PR -->
**What changed?**

Active-active domain failovers (changing the region to cluster map)
increment the failover version of the updated entries. Rest of the
entries are untouched. This causes a problem when an ongoing workflow is
attempted to be started after failover.

Example scenario:
- Domain's initial state
```
ActiveClustersByRegion: [
    phx: {staging_phx 2} 
    dca: {staging_dca 0}
]
```

- A workflow with id `cron.phx` is running with version 2. It's active
on staging_phx cluster.
- Domain is failed over from PHX to DCA by operator/automation. The map
in DB now looks like this:
```
ActiveClustersByRegion: [
    phx: {staging_dca 10} # version is incremented from 2 to 10.
    dca: {staging_dca 0}
]
```

- A `StartWorkflow(cron.phx)` request is made by a client.
- Call arrives to PHX frontend.
- It checks the `ActiveClustersByRegion` of the domain and decides to
forward to DCA frontend.
- DCA frontend receives the request and makes corresponding request to
DCA history.
- The history engine responsible from `cron.phx` shard processes the
request.
- It gets workflow already started error and checks the new mutable
state version with the existing one in DB.
- Existing mutable state in DB (which was replicated from PHX cluster)
has version 2.
- New mutable state in memory has version 0 which is the version DCA
uses.
- New version is less than previous version so it returns domain not
active error
[ref](https://github.com/cadence-workflow/cadence/blob/147489a7e507a04eade6594854234396daebcd8f/service/history/engine/engineimpl/start_workflow_execution.go#L253)


The problem is caused by multiple entries in the
`ActiveClustersByRegion` map pointing to the same cluster but having
different versions.

One way to prevent getting into this state is to increment failover
versions of all entries pointing to the same cluster when updating
ActiveClustersByRegion.
Before:
```
ActiveClustersByRegion: [
    phx: {staging_phx 2} 
    dca: {staging_dca 0}
]
```

After:
```
ActiveClustersByRegion: [
    phx: {staging_dca 10}  # incremented from 2 to 10 so it points to dca cluster
    dca: {staging_dca 10} #  incremented from 0 to 10 so it's greater than or equal to the entry that was updated (above)
]
```

**Validation Change**
Another change in this PR is to restrict what kind of updates are
allowed for `ActiveClustersByRegion` map. It doesn't make sense to
support multiple hops and we should prevent cycles.

For example below map contains multiple hops and will not be allowed. 
```
ActiveClustersByRegion: [
    phx: {staging_dca} 
    dca: {staging_klm}
    klm: {staging_klm}
]
```


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
- unit tests
- new simulation