Fix virtual slice's get task method (#7249)
<!-- Describe what has changed in this PR -->
**What changed?**
Update virtual slice's get task method to return partial result

<!-- Tell your future self why have you made these changes -->
**Why?**
We're seeing a latency increase when there is a deployment of
cadence-history. And also the ack level of some shards are not updated
for several hours. This means that we have some zombie tasks that is
loaded into memory but not being scheduled. After reading the code, I
found that it could happen when we only load partial results from a
virtual slice. This is to fix this issue.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
Task processing might be broken

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**