Fix a bug that cause non-default virtual queue to be orphaned (#7263)
<!-- Describe what has changed in this PR -->
**What changed?**
Fix a bug that cause non-default virtual queue to be orphaned

<!-- Tell your future self why have you made these changes -->
**Why?**
When virtual_queue component's UpdateAndGetState method is called, it
can remove a slice pointed by sliceToRead Pointer. This is not expected,
but what happens is that in that virtual slice component, the
nextTaskKey is equals to the ExclusiveMaxTaskKey of that range, but
nextPageToken is not empty. So HasMoreTasks method still returns true,
but UpdateAndGetState method of virtual slice component updates
InclusiveMinTaskKey to be the same as nextTaskKey which is the same as
ExclusiveMaxTaskKey.

This doesn't affect root virtual queue, because when new virtual slices
are merged into root queue, sliceToRead pointer is reset to be non-nil.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests and load tests in dev2

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**