Improve conflict resolution for active-active (#7265)
<!-- Describe what has changed in this PR -->
**What changed?**
Improve conflict resolution for active-active and add a new simulation
test to verify it

The simulation test starts 2 workflows in different clusters with the
same id concurrently. We expect the one started earlier to be terminated
and the other one to complete.

<!-- Tell your future self why have you made these changes -->
**Why?**
For workflows from an active-active domain's conflict resolution, we
cannot simply compare their failover version to determine which one is
newer if they have different active selection policies (failover
groups), because the failover version is associated with an active
selection policy (failover group).

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests & simulation tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**