Test active-active behavior with concurrent SignalWithStart requests (#7270)
<!-- Describe what has changed in this PR -->
**What changed?**

Tests conflict resolution when the same workflowID is used to create
workflows in two different clusters with an Active-Active setup in the
domain.

<!-- Tell your future self why have you made these changes -->
**Why?**

Cadence does not support two workflows existing with the same workflow
ID. When a domain is configured active-active it can be possible for a
user to make two simultaneous requests to two different clusters with
the same workflow ID. When this happens, both workflows will run
temporarily until the clusters have replicated their data to the other,
at which point the conflict will have to be resolved.

This test is a follow up to
https://github.com/cadence-workflow/cadence/pull/7265, testing the
scenario when SignalWithStart is used (rather than StartWorkflow).

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

Locally + against dev canary domains. 

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

N/A

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A