Add documentation to the simulation tests (#7271)
<!-- Describe what has changed in this PR -->
**What changed?**

Updated the documentation in the simulation folder. 
Updated each `run.sh` script to:
- support named parameters, allowing users to only specify the fields
they need
- support a `dockerfile-suffix`, allowing users to run the
matching/history sims locally
- require a scenario file

<!-- Tell your future self why have you made these changes -->
**Why?**

My hope for the documentation is that it will increase usage of the
simulation folder by decreasing the barrier to running simulations.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

Manually - needs to also be tested in CI/CD still. 

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

CI/CD could stop working if the simulations can't be run in Canary. 

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Not sure - is it worth writing something about how to use simulations?