feat: pull in IDL changes, pull in failover-endpoint changes and active-active clusterConfig refactor (#7285)
<!-- Describe what has changed in this PR -->
**What changed?**

This pulls in the tip of master branch of the IDL and stubs out the
associated endpoints brought in as a result:

- A new purpose-specific `Failover` endpoint which is intended to
replace the use of UpdateDomain
- A set of refactoring for the ActiveActive cluster

The failover endpoint is *not* implemented in this diff, only stubbed
out. The mappers, tests and so on are to follow.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**