feat [shard-distributor]: Removed custom storage registration logic (#7274)
<!-- Describe what has changed in this PR -->
**What changed?**
Removed custom logic to store and discover storage backends, and moved
to using fx


<!-- Tell your future self why have you made these changes -->
**Why?**
The logic was not nessesary and didn't acheive the goal of avoiding a
dependencty.
We want to be able to add things to the storage fx modules which was not
possible with this solution.


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit tests


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: Jakob Haahr Taankvist <jht@uber.com>