chore(replication-cache): make replication cache generic (BoundedAckCache) (#7287)
<!-- Describe what has changed in this PR -->
**What changed?**
Made replication cache generic, moved it to common/cache as
BoundedAckCache. This cache does not evict on add and rejects adds if
full. It expects a sequential consumption and eviction is done by acking
it with a sequenceID.

<!-- Tell your future self why have you made these changes -->
**Why?**
Make the cache reusable.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit tests. It will be further tested on in-memory replication by host
implementation.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
Currently not used by default. May affect users that use in-memory
replication but that's not recommended as it currently has risk of
causing OOM kill.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: fimanishi <fimanishi@gmail.com>