feat: Add AttributeScope to internal ActiveClusters type, support mapping from proto and thrift (#7288)
<!-- Describe what has changed in this PR -->
**What changed?**

Add AttributeScope to internal ActiveClusters type, support mapping from
proto and thrift

<!-- Tell your future self why have you made these changes -->
**Why?**

Active-Active is introducing support for ClusterAttributes - a way to
define which cluster a subset of workflows (tagged with the attribute)
get run in.

This adds support for mapping the new external field to our internal
types to enable active-active domains to specify a ClusterAttribute.
This is the bare minimum (I believe) that will allow us to parallelize
the work to add this support across the product.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

Just with unit tests. 

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

Should not affect existing behavior. 

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A