chore(active-active): Remove ExternalEntityProvider (#7292)
<!-- Describe what has changed in this PR -->
**What changed?**
Remove ExternalEntityProvider and related code. The component was
designed to be a watcher of external entities to update failover
versions of external entities. We don't need this component anymore.

<!-- Tell your future self why have you made these changes -->
**Why?**
ExternalEntityProvider is not needed because we redesign active-active

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
unit tests & replication simulation tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**