refactor: Support ClusterAttributes from existing ActiveActive helpers (#7291)
<!-- Describe what has changed in this PR -->
**What changed?**

Adds helper functions to work with ClusterAttributes for active-active
domains. Refactors the data type to use values instead of pointers in
each map.

<!-- Tell your future self why have you made these changes -->
**Why?**

These helper functions are intended to replace direct references to the
ActiveClusters map, and make it significantly easier to interact with
Active-Active domains.

Updating the data type makes the ClusterAttribute maps consistent with
the ClusterToRegion map implementation. Ensuring that everything within
the map has a concrete value reduces the amount of `nil` handling we
need to do, and removes the footgun of accidentally referencing a type
on a nil value.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

Just unit tests so far. Once we switch on ClusterAttributes in some
clusters we will ensure that sim tests/canary tests pass with the new
implementation.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

N/A

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A