fix(histograms): depend only on the histogram config, not all config (#7310)
Followup on #7298 to be a bit easier to set up.
This is pretty minor, but helps us internally (and possibly others) by
not requiring the "main" config object (which we do not use internally,
at least in our main service binary).

Anyone using the standard fx startup should be unaffected, as it now
also provides the finer-grained histogram config.
Otherwise, if you get an error about `metrics.HistogramMigration` not
being provided: just `fx.Provide(func(...) metrics.HistogramMigration {
... })` it however you like.

And, last but not least: apparently automerge used an old commit message
for #7298. Apologies about that, check the PR for an updated and much
more descriptive description.

Signed-off-by: Steven L <imgroxx@gmail.com>