feat:[ShardDistributor] add external assignment in executor library (#7312)
<!-- Describe what has changed in this PR -->
**What changed?**
Adding a method to assign shards to an executor, this is communicating
the assignment to the SD using heartbeat, this is the first step for the
migration.

<!-- Tell your future self why have you made these changes -->
**Why?**
This is required during the migration to use the ShardDistributor, such
that we can assign a specific shard to an executor and keep using the
current sharding logic

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
integration in shard-distribution-canary with a dedicated module that
execute the assignment

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**
no risk, since the executor client is not integrated with any service
yet.

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>