feat:[Shard distributor] Include migration mode in the heartbeat (#7316)
<!-- Describe what has changed in this PR -->
**What changed?**
Extending the heartbeat response to include the migration mode

<!-- Tell your future self why have you made these changes -->
**Why?**
We need to know if the shard executor should apply the new assignment
received as response to the heartbeat

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
running make pr

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
the endpoint is only used from the executor that are in development mode

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>