feat(Shard Distributor): In-memory cache for shard to executor, and remove shard to executor mapping in ETCD (#7289)
<!-- Describe what has changed in this PR -->
**What changed?**
We now have an in-memory cache for each of the name spaces on each shard
distributor instance. The cache will be updated when things change in
the database, using ETCD watches.

Also separated the ETCD store into more modules.

<!-- Tell your future self why have you made these changes -->
**Why?**
Before we attempted to push all the shard-to-executor mappings directly
to the ETCD in a single transaction, this didn't work due to the number
of transactions being way too big for ETCD.


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit tests and local runs 

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
Currently the subscriptions to the database produces too many events,
will address this in the next PR

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: Jakob Haahr Taankvist <jht@uber.com>