refactor: Remove panics in proto/thrift to internal type mapping functions (#7258)
<!-- Describe what has changed in this PR -->
**What changed?**

Removes `panic()` from enum mappers in types/mapper/...

<!-- Tell your future self why have you made these changes -->
**Why?**

It is [idiomatic in
go](https://github.com/uber-go/guide/blob/master/style.md#dont-panic) to
[return an error value](https://gobyexample.com/panic) when something
expected goes wrong.
In general panics should be reserved for situations which are
unrecoverable for the entire process, and killing the process and
restarting it is the expected behavior.
Even in these scenarios, libraries or portions of the call stack that
are not at the request level should favor returning errors to the caller
and allowing the caller to decide whether or not the runtime should
panic - deferring the decision to the last moment (e.g a panic/recover
handler at the request level.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

There are three scenarios for each of our mapping functions:
- An expected invalid value is passed (`nil` or `PROTO_ENUM_INVALID`)
- A valid value is passed
- An invalid value/out of bounds value is passed

With the existing mapping round trip tests scenarios 1 and 2 were
tested, but scenario 3 was not.
This PR adds tests for the ToXXX and FromXXX mappers as well to ensure
that there aren't strange behaviors that are covered up by the full
round trip.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

This could be a breaking change to our API for some users. If, for
example, a consumer is passing an invalid enum value to a specific RPC
and then relying on an Internal_Server_Error with a specific panic
message to do some business logic, they will unfortunately not be happy.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A