test: Ensure activeclustername is set for active-active domains (#7314)
<!-- Describe what has changed in this PR -->
**What changed?**

Updates the simulation tests to support specifying ClusterAttributes as
part of the Register and Update domain requests. Requires
ActiveClusterName to be specified for all global domains.

<!-- Tell your future self why have you made these changes -->
**Why?**

To support ClusterAttributeScopes as the new method for dividing
active-active domains into sub-domains the simulation tests must be
updated to validate that this works. They have been added in a backwards
compatible way, with an eventual switch over to be done at a later date
(along with cleanup of the existing region concept).

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

Who tests the testers???

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

N/A

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A