fix: correct cluster attributes being set when both legacy and new are passed in  (#7330)
<!-- Describe what has changed in this PR -->
**What changed?**

in some of the integration-tests, they were not failing over a domain's
cluster-attributes correctly. This was due to a bug in the
merging-cluster-attributes code which didn't correctly flag the
attributes as having changed when they should have. The whole section
needs both better validation and more test coverage. I've a few followup
PRs for this.

As an aside, there actual input-data is quite weird, so I think there's
both some test setup problems which I've not identified and possibly a
strange/possibly wrong assumption in the previous attributesByRegion
code. But since it's going away maybe that doesn't matter.


<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>