chore: Move multi cluster script to scripts (#7327)
<!-- Describe what has changed in this PR -->
**What changed?**

Moved the `test_multicluster_domain_workflow.sh` to scripts. 

<!-- Tell your future self why have you made these changes -->
**Why?**

Cleanup the main dir and move it to the appropriate place. This should
be migrated to a simulation/canary test once headers are added to return
the cluster that serviced a request.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

N/A

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

N/A

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A