feat: Adding metrics to track workflow access (#7331)
<!-- Describe what has changed in this PR -->
**What changed?**
This PR adds a new metric to track domain retention days for workflows
accessing the `GetWorkflowExecutionHistory` endpoint, tagged with
workflow ID, domain, and workflow close status.

<!-- Tell your future self why have you made these changes -->
**Why?**
This enables us to monitor if users are using the given retention days
for closed workflows. And decrease retention days if they are not
utilizing it properly.
This will allow us to see the real workflow access pattern.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Local testing with Prometheus

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: Gaziza Yestemirova <gaziza@uber.com>