docs: Convert slack links to CNCF, add link to contributing guide (#7337)
<!-- Describe what has changed in this PR -->
**What changed?**

Converts links from our old slack instance to the CNCF community
inviter.
Adds a link to our shared contributing guide to the contributing readme.

<!-- Tell your future self why have you made these changes -->
**Why?**

The goal of the contribution guide is to move all common guides and
processes between our repositories to a shared place – reducing
conflicting information and streamlining the process to contribute to
cadence.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

N/A

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

N/A

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A